requirements.txt
torch
torchvision
torchaudio
roboflow
python-dotenv
opencv-python
mediapipe
pycocotools
pillow
matplotlib
numpy
